05.13

+++ final/lib/applicationState.dart
@@ -21,6 +21,7 @@ import 'package:flutter/material.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:provider/provider.dart';
 import 'package:shrine/app.dart';
+import 'package:shrine/profile.dart';
 
 import 'firebase_options.dart';
 import 'product.dart';
@@ -44,25 +45,16 @@ class ApplicationState extends ChangeNotifier {
             .listen((snapshot) {
           _productList = [];
           for (final document in snapshot.docs) {
-            // if(document.data()['creationTimestamp'] == null){
-            //   FirebaseFirestore.instance
-            //       .collection('product')
-            //       .doc(document.id)
-            //       .update({
-            //     'creationTimestamp': FieldValue.serverTimestamp(),
-            //     'updateTimestamp': FieldValue.serverTimestamp(),
-            //   }) ;
-            // }
             _productList.add(
               ProductElement(
                 docId: document.id,
-                userId: document.data()['uid'] as String,
+                userId: document.data()['userId'] as String,
                 likeNum: document.data()['likeNum'] as int,
                 name: document.data()['name'] as String,
                 price: document.data()['price'] as int,
                 description: document.data()['description'] as String,
-                creationTime: document.data()['creationTimestamp'],
-                updateTime: document.data()[''],
+                creationTime: document.data()['creationTimestamp'] as Timestamp,
+                updateTime: document.data()['updateTimestamp'] as Timestamp,
                 path: document.data()['path'] as String,
               ),
             );
@@ -71,9 +63,10 @@ class ApplicationState extends ChangeNotifier {
         });
       } else {
         _loginState = ApplicationLoginState.loggedOut;
+        _user = UserElement() ;
         _productList = [];
         _productListSubscription?.cancel();
-        signOut() ;
+        signOut();
       }
       notifyListeners();
     });
@@ -81,13 +74,17 @@ class ApplicationState extends ChangeNotifier {
 
   ApplicationLoginState _loginState = ApplicationLoginState.loggedOut;
   ApplicationLoginState get loginState => _loginState;
+  UserElement _user = UserElement();
+  UserElement get user => _user;
 
   Future<UserCredential> signInWithGoogle() async {
+    _loginState = ApplicationLoginState.googleLogin;
     // Trigger the authentication flow
     final GoogleSignInAccount? googleUser = await GoogleSignIn().signIn();
 
     // Obtain the auth details from the request
-    final GoogleSignInAuthentication? googleAuth = await googleUser?.authentication;
+    final GoogleSignInAuthentication? googleAuth = await googleUser
+        ?.authentication;
 
     // Create a new credential
     final credential = GoogleAuthProvider.credential(
@@ -101,26 +98,24 @@ class ApplicationState extends ChangeNotifier {
 
   Future<void> startGoogleLoginFlow() async {
     _loginState = ApplicationLoginState.googleLogin;
-    UserCredential userCredential = await signInWithGoogle() ;
+    UserCredential userCredential = await signInWithGoogle();
     final user = userCredential.user;
     if (user != null) {
-      for (final providerProfile in user.providerData) {
-        final provider = providerProfile.providerId;
-        final uid = providerProfile.uid;
-        final name = providerProfile.displayName;
-        final emailAddress = providerProfile.email;
-        final profilePhoto = providerProfile.photoURL;
-      }
+      _user.userId = user.providerData[0].uid!;
+      _user.email = user.providerData[0].email!;
+      _user.path = user.providerData[0].photoURL!;
     }
   }
 
   Future<void> startGuestLoginFlow() async {
     _loginState = ApplicationLoginState.guestLogin;
-    try {
-      UserCredential userCredential = await FirebaseAuth.instance.signInAnonymously();
-      User? user = userCredential.user;
-    } catch (e) {
-      print(e.toString());
+    UserCredential userCredential = await FirebaseAuth.instance.signInAnonymously();
+    final user = userCredential.user;
+    if (user != null) {
+        _user.userId = user.uid!;
+        _user.email = 'Anonymous' ;
+        getDefaultImage()
+            .then((value) => _user.path = value) ;
     }
   }
 
@@ -139,7 +134,8 @@ class ApplicationState extends ChangeNotifier {
   List<ProductElement> _productList = [];
   List<ProductElement> get productList => _productList;
 
-  Future<DocumentReference> addProduct(String name, int price, String description, String path) {
+  Future<DocumentReference> addProduct(String name, int price,
+      String description, String path) {
     if (loginState == ApplicationLoginState.loggedOut) {
       throw Exception('Must be logged in');
     }
@@ -147,18 +143,19 @@ class ApplicationState extends ChangeNotifier {
     return FirebaseFirestore.instance
         .collection('product')
         .add({
-      'uid': FirebaseAuth.instance.currentUser!.uid,
+      'userId': FirebaseAuth.instance.currentUser!.uid,
       'likeNum': 0,
       'name': name,
       'price': price,
       'description': description,
       'path': path,
-      'timestamp' : FieldValue.serverTimestamp(),
-      'updateTimestamp' : FieldValue.serverTimestamp(),
+      'creationTimestamp': FieldValue.serverTimestamp(),
+      'updateTimestamp': FieldValue.serverTimestamp(),
     });
   }
 
-  Future<void> editProduct(String docId, String name, int price, String description, String path) {
+  Future<void> editProduct(String docId, String name, int price,
+      String description, String path) {
     if (loginState == ApplicationLoginState.loggedOut) {
       throw Exception('Must be logged in');
     }
@@ -171,13 +168,71 @@ class ApplicationState extends ChangeNotifier {
       'price': price,
       'description': description,
       'path': path,
-      'updateTimestamp' : FieldValue.serverTimestamp(),
+      'updateTimestamp': FieldValue.serverTimestamp(),
     });
   }
 
   Future<void> deleteProductInServer(String docId) async {
     FirebaseFirestore.instance.collection('product').doc(docId).delete();
+
+    FirebaseFirestore.instance
+        .collection('likeProduct')
+        .where('productId', isEqualTo: docId)
+        .where('userId', isEqualTo: FirebaseAuth.instance.currentUser!.uid)
+        .get()
+        .then((snapshot) => {
+      for (final document in snapshot.docs){
+        FirebaseFirestore.instance.collection('likeProduct')
+            .doc(document.id).delete()
+      }}) ;
     notifyListeners();
   }
-}
 
+  Future<int> alreadyLikeCheck(String docId) {
+    if (loginState == ApplicationLoginState.loggedOut) {
+      throw Exception('Must be logged in');
+    }
+    bool _isExist;
+    return FirebaseFirestore.instance
+        .collection('likeProduct')
+        .where('productId', isEqualTo: docId)
+        .where('userId', isEqualTo: FirebaseAuth.instance.currentUser!.uid)
+        .get()
+        .then((value) => value.size);
+  }
+
+  Future<void> increaseLike(String docId) async {
+    if (loginState == ApplicationLoginState.loggedOut) {
+      throw Exception('Must be logged in');
+    }
+    FirebaseFirestore.instance
+        .collection('product')
+        .doc(docId)
+        .update({'likeNum': FieldValue.increment(1)});
+
+    FirebaseFirestore.instance
+        .collection('likeProduct')
+        .add({
+      'userId': FirebaseAuth.instance.currentUser!.uid,
+      'productId': docId,
+      'timestamp' : FieldValue.serverTimestamp(),
+    });
+  }
+
+  String getCurrentUserId() {
+    if (loginState == ApplicationLoginState.loggedOut) {
+      throw Exception('Must be logged in');
+    }
+    return FirebaseAuth.instance.currentUser!.uid ;
+  }
+
+  Future<String> getDefaultImage() {
+    return FirebaseFirestore.instance
+        .collection('initialize')
+        .doc('productDefaultImage')
+        .get().then(
+          (doc) => doc['logo'] as String,
+      onError: (e) => print("Error completing: $e"),
+    );
+  }
+}
